Rename files to follow ruby naming conventions
If a class is named NullLogger, it’s file would be named null_logger.rb
