+++ lib/rack/common_logger.rb
