+++ lib/rack/conditional_get.rb
