+++ lib/rack/null_logger.rb
