+++ lib/rack/show_exceptions.rb
